Remove null values when performing aggregate. (those null value do not get display in scatter plots) Refactor charts.js into two files Move aggregation and formatting function to utils.js Cleanup
