+++ reporting_app/__init__.py
@@ -109,7 +109,7 @@ def change_password():
 @app.route('/runs/<view_type>')
 @flask_login.login_required
 def runs_report(view_type):
-    ajax_call = {'func_name': 'merge_multi_sources', 'merge_on': 'run_id'}
+    ajax_call = {'func_name': 'dt_merge_multi_sources', 'merge_on': 'run_id'}
 
     if view_type == 'all':
         ajax_call['api_urls'] = [
@@ -220,7 +220,7 @@ def project_reports():
             'All projects list',
             'projects',
             ajax_call={
-                'func_name': 'merge_multi_sources',
+                'func_name': 'dt_merge_multi_sources',
                 'api_urls': [
                     util.construct_url('projects', max_results=10000),
                     util.construct_url('lims/project_info', match={'project_status': 'all'}),
@@ -235,7 +235,7 @@ def project_reports():
 @flask_login.login_required
 def report_samples(view_type):
     six_months_ago = util.now() - datetime.timedelta(days=182)
-    ajax_call = {'func_name': 'merge_multi_sources_keep_first', 'merge_on': 'sample_id'}
+    ajax_call = {'func_name': 'dt_merge_multi_sources_keep_first', 'merge_on': 'sample_id'}
 
     if view_type == 'all':
         title = 'All samples'
@@ -290,7 +290,7 @@ def report_project(project_ids):
     if len(id_list) > 1:
         project_status_call = {
             'ajax_call': {
-                'func_name': 'merge_multi_sources',
+                'func_name': 'dt_merge_multi_sources',
                 'merge_on': 'project_id',
                 'api_urls': [
                     util.construct_url('lims/project_status', match={'project_id': i, 'project_status': 'all'})
@@ -300,7 +300,7 @@ def report_project(project_ids):
         }
         plate_status_call = {
             'ajax_call': {
-                'func_name': 'merge_multi_sources',
+                'func_name': 'dt_merge_multi_sources',
                 'merge_on': 'plate_id',
                 'api_urls': [
                     util.construct_url('lims/plate_status', match={'project_id': i, 'project_status': 'all'})
@@ -355,7 +355,7 @@ def report_project(project_ids):
                 'Bioinformatics report for ' + project_ids,
                 'samples',
                 ajax_call={
-                    'func_name': 'merge_multi_sources',
+                    'func_name': 'dt_merge_multi_sources',
                     'api_urls': bioinformatics_urls,
                     'merge_on': 'sample_id'
                 },
@@ -380,7 +380,7 @@ def report_sample(sample_id):
                 'Bioinformatics report for ' + sample_id,
                 'samples',
                 ajax_call={
-                    'func_name': 'merge_multi_sources',
+                    'func_name': 'dt_merge_multi_sources',
                     'merge_on': 'sample_id',
                     'api_urls': [
                         util.construct_url('samples', where={'sample_id': sample_id}),
@@ -566,9 +566,9 @@ def species_page(species):
     )
 
 
-@app.route('/charts/<view_type>')
+@app.route('/charts/seq/<view_type>')
 @flask_login.login_required
-def plotting_report(view_type):
+def sequencing_charts(view_type):
     if view_type == 'last_month':
         time_ago = util.now() - datetime.timedelta(days=30)
     elif view_type == 'last_3_months':
@@ -578,12 +578,12 @@ def plotting_report(view_type):
     else:
         fl.abort(404)
         return None
-
+    time_str = time_ago.strftime(settings.DATE_FORMAT)
     return render_template(
         'charts.html',
         api_urls=[
-            util.construct_url('lanes', max_results=10000, where={'_created': {'$gte': time_ago.strftime(settings.DATE_FORMAT)}}),
-            util.construct_url('lims/run_status', createddate=time_ago.strftime(settings.DATE_FORMAT)),
+            util.construct_url('lanes', max_results=10000, where={'_created': {'$gte': time_str}}),
+            util.construct_url('lims/run_status', createddate=time_str),
         ],
         ajax_token=util.get_token(),
         merge_on='run_id',
@@ -593,22 +593,11 @@ def plotting_report(view_type):
     )
 
 
-@app.route('/tat_chart/')
+@app.route('/charts/tat')
 @flask_login.login_required
-def tat_chart_page(view_type):
-    time_ago = None
-    if view_type == 'last_month':
-        time_ago = util.now() - datetime.timedelta(days=30)
-    if view_type == 'last_3_months':
-        time_ago = util.now() - datetime.timedelta(days=182)
-    elif view_type == 'last_12_months':
-        time_ago = util.now() - datetime.timedelta(days=365)
-    else:
-        fl.abort(404)
-        return None
-
+def tat_charts():
     return render_template(
         'tat_charts.html',
-        api_url=util.construct_url('lims/sample_status', match={'createddate': time_ago.strftime(settings.DATE_FORMAT), 'project_status': 'open'}),
+        api_url=util.construct_url('lims/sample_status', match={'project_status': 'all'}),
         ajax_token=util.get_token()
     )
