+++ reporting_app/static/datatable.js
@@ -1,97 +1,21 @@
-////helper function to retrieve a function by name
-//var get_function = function (fn_name){
-//    var fn = window[fn_name];
-//    if(typeof fn === 'function') {
-//        return fn;
-//    }else{
-//        return null;
-//    }
-//}
+// datatables specific functions for merge_multi_sources and merge_multi_sources_keep_first
+var dt_merge_multi_sources = function(dt_config){
+    return merge_multi_sources(
+        dt_config.ajax_call.api_urls,
+        dt_config.token,
+        dt_config.ajax_call.merge_on,
+        dt_config.ajax_call.merged_properties
+    );
+}
 
-
-//// merge several arrays of object based on given property
-//// see http://stackoverflow.com/questions/38053193/merge-2-json-array-objects-based-on-a-common-property
-//var merge_on = function (list_of_array, key) {
-//    var r = [],
-//        hash = Object.create(null);
-//
-//    list_of_array.forEach(function (a) {
-//        a.forEach(function (o) {
-//            if (!hash[o[key]]) {
-//                hash[o[key]] = {};
-//                r.push(hash[o[key]]);
-//            }
-//            Object.keys(o).forEach(function (k) {
-//                hash[o[key]][k] = o[k];
-//            });
-//        });
-//    });
-//    return r;
-//}
-//
-//// merge several arrays of object based on given property but only keep object present in the first array
-//// the other ones can add properties but no new objects
-//var merge_on_keep_first = function (list_of_array, key) {
-//    var r = [],
-//        hash = Object.create(null);
-//
-//    list_of_array[0].forEach(function (o) {
-//        hash[o[key]] = {};
-//        r.push(hash[o[key]]);
-//    });
-//    list_of_array.forEach(function (a) {
-//        a.forEach(function (o) {
-//            if (hash[o[key]]) {
-//                Object.keys(o).forEach(function (k) {
-//                    hash[o[key]][k] = o[k];
-//                });
-//            }
-//        });
-//    });
-//    return r;
-//}
-
-// send multiple ajax queries then merge the results based on dt_config.merge_on
-// dt_config is expected to contain:
-// ajax_call.func_name: merge_multi_sources,
-// ajax_call.api_urls: [url1, url2, ...]
-// ajax_call.merge_on: property_name
-// token: the token used for authentication
-//var _merge_multi_sources = function(dt_config, merge_func){
-//    return function(data, callback, settings){
-//        var calls = dt_config.ajax_call.api_urls.map( function(api_url){
-//            return $.ajax({
-//                url: api_url,
-//                headers: {'Authorization':  dt_config.token },
-//                dataType: 'json',
-//                async: true,
-//            });
-//        });
-//
-//        // pass an array of deferred calls and apply with Function.prototype.apply
-//        // see http://stackoverflow.com/questions/5627284/pass-in-an-array-of-deferreds-to-when
-//        $.when.apply($, calls).then(function () {
-//            // Use 'arguments' to get all the responses as an array-like object.
-//            // then extract the data field
-//            var data_array = _.map(arguments, function(response){return response[0].data});
-//            var result = merge_func(data_array, dt_config.ajax_call.merge_on);
-//            callback({
-//                recordsTotal: result.length,
-//                recordsFiltered: result.length,
-//                data: result
-//            });
-//        });
-//    }
-//}
-//
-//var merge_multi_sources = function(dt_config){
-//    return _merge_multi_sources(dt_config, merge_on);
-//}
-//
-//var merge_multi_sources_keep_first = function(dt_config){
-//    return _merge_multi_sources(dt_config, merge_on_keep_first);
-//}
-
+var dt_merge_multi_sources_keep_first = function(dt_config){
+    return merge_multi_sources_keep_first(
+        dt_config.ajax_call.api_urls,
+        dt_config.token,
+        dt_config.ajax_call.merge_on,
+        dt_config.ajax_call.merged_properties
+    );
+}
 
 var required_yields = function(dt_config) {
     return function(data, callback, settings) {
@@ -295,12 +219,7 @@ var configure_dt = function(dt_config) {
     }
     if (dt_config.ajax_call){
         // retrieve the function generating the ajax calls by name and call it with the config
-        ajax_call = get_function(dt_config.ajax_call.func_name)(
-            dt_config.ajax_call.api_urls,
-            dt_config.token,
-            dt_config.ajax_call.merge_on,
-            dt_config.ajax_call.merged_properties
-        );
+        ajax_call = get_function(dt_config.ajax_call.func_name)(dt_config);
     }
     return {
         'dt_config': dt_config,
