+++ reporting_app/templates/base.html
@@ -91,9 +91,12 @@
                     <li class="dropdown">
                         <a href="#" class="dropdown-toggle" data-toggle="dropdown">Charts<b class="caret"></b></a>
                         <ul class="dropdown-menu">
-                            <li><a href="/charts/last_month">30 days</a></li>
-                            <li><a href="/charts/last_3_months">Last 3 month</a></li>
-                            <li><a href="/charts/last_12_months">Last 12 months</a></li>
+                            <h6 class="dropdown-header">Seq metrics</h6>
+                            <li><a href="/charts/seq/last_month">Last 30 days</a></li>
+                            <li><a href="/charts/seq/last_3_months">Last 3 months</a></li>
+                            <li><a href="/charts/seq/last_12_months">Last 12 months</a></li>
+                            <h6 class="dropdown-header">TAT metric</h6>
+                            <li><a href="/charts/tat">Since start</a></li>
                         </ul>
                     </li>
                     <li class="dropdown">
