+++ reporting_app/templates/charts.html
@@ -3,14 +3,15 @@
 {% block head %}
     {{ super() }}
     <link rel="stylesheet" type="text/css" href="//cdnjs.cloudflare.com/ajax/libs/font-awesome/4.4.0/css/font-awesome.min.css"/>
-    <script src="https://code.highcharts.com/highcharts.js"></script>
-    <script src="https://code.highcharts.com/highcharts-more.js"></script>
-    <script src="https://code.highcharts.com/modules/exporting.js"></script>
-    <script src="//rawgithub.com/phpepe/highcharts-regression/master/highcharts-regression.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.13.0/moment.min.js"></script>
-    <script src="https://cdn.jsdelivr.net/lodash/4.17.4/lodash.min.js"></script>
-    <script src="http://cdnjs.cloudflare.com/ajax/libs/mathjs/3.10.0/math.min.js"></script>
-    <script src="{{ url_for('static', filename='charts.js') }}"></script>
+    <script type="text/javascript" charset="utf8" src="https://code.highcharts.com/7.1.2/highcharts.js"></script>
+    <script type="text/javascript" charset="utf8" src="https://code.highcharts.com/7.1.2/highcharts-more.js"></script>
+    <script type="text/javascript" charset="utf8" src="https://code.highcharts.com/7.1.2/modules/exporting.js"></script>
+    <script type="text/javascript" charset="utf8" src="//rawgithub.com/streamlinesocial/highcharts-regression/b84143a4e78a8e128374fe485c711b09a1062750/highcharts-regression.js"></script>
+    <script type="text/javascript" charset="utf8" src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.13.0/moment.min.js"></script>
+    <script type="text/javascript" charset="utf8" src="https://cdn.jsdelivr.net/lodash/4.17.4/lodash.min.js"></script>
+    <script type="text/javascript" charset="utf8" src="https://cdnjs.cloudflare.com/ajax/libs/mathjs/5.10.3/math.min.js" crossorigin="anonymous"></script>
+
+    <script src="{{ url_for('static', filename='seq_charts.js') }}"></script>
     <script src="{{ url_for('static', filename='utils.js') }}"></script>
 {% endblock %}
