+++ tests/js/run_tests.html
@@ -12,13 +12,15 @@
     <script type="text/javascript" charset="utf8" src="https://code.jquery.com/jquery-1.10.2.min.js"></script>
     <script type="text/javascript" charset="utf8" src="https://cdnjs.cloudflare.com/ajax/libs/moment.js/2.13.0/moment.min.js"></script>
     <script type="text/javascript" charset="utf8" src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.6/js/bootstrap.min.js" crossorigin="anonymous"></script>
+    <script type="text/javascript" charset="utf8" src="https://cdnjs.cloudflare.com/ajax/libs/mathjs/5.10.3/math.min.js" crossorigin="anonymous"></script>
     <script src="https://code.highcharts.com/highcharts.js"></script>
     <script src="../../reporting_app/static/humanize.js"></script>
 
     <!-- things to test -->
     <script src="../../reporting_app/static/column_formatter.js"></script>
     <script src="../../reporting_app/static/datatable.js"></script>
-    <script src="../../reporting_app/static/charts.js"></script>
+    <script src="../../reporting_app/static/seq_charts.js"></script>
+    <script src="../../reporting_app/static/tat_charts.js"></script>
     <script src="../../reporting_app/static/library_preps.js"></script>
     <script src="../../reporting_app/static/utils.js"></script>
